+++ sopa/_sdata.py
@@ -54,7 +54,7 @@ def get_boundaries(
     if not warn:
         raise ValueError(error_message)
 
-    log.warn(error_message)
+    log.warning(error_message)
     return (None, None) if return_key else None