+++ sopa/annotation/tangram/run.py
@@ -64,7 +64,7 @@ class MultiLevelAnnotation:
         self.ad_sc = ad_sc
 
         if not self.ad_sp.uns.get(SopaKeys.UNS_KEY, {}).get(SopaKeys.UNS_HAS_TRANSCRIPTS, False):
-            log.warn(
+            log.warning(
                 "The values in adata.X doesn't seem to be transcript count, which may create unexpected behaviors when running Tangram."
             )
 
@@ -212,7 +212,7 @@ class MultiLevelAnnotation:
             )
 
         if indices_sp is not None and len(indices_sp) == 0:
-            log.warn("No cell annotated in the upper level...")
+            log.warning("No cell annotated in the upper level...")
             return
 
         indices_sp = self.ad_sp.obs_names if indices_sp is None else indices_sp