+++ sopa/io/explorer/points.py
@@ -4,15 +4,12 @@ import logging
 from math import ceil
 from pathlib import Path
 
-import dask
+import dask.dataframe as dd
 import numpy as np
 import zarr
 
 from ._constants import ExplorerConstants, FileNames
 from .utils import explorer_file_path
-
-dask.config.set({"dataframe.query-planning": False})
-import dask.dataframe as dd  # noqa
 
 log = logging.getLogger(__name__)
 
@@ -54,7 +51,7 @@ def write_transcripts(
     location = np.concatenate([location, np.zeros((num_transcripts, 1))], axis=1)
 
     if location.min() < 0:
-        log.warn("Some transcripts are located outside of the image (pixels < 0)")
+        log.warning("Some transcripts are located outside of the image (pixels < 0)")
     log.info(f"Writing {len(df)} transcripts")
 
     xmax, ymax = location[:, :2].max(axis=0)